==> 02.GroupDF.ipynb <==
import pandas as pd

df = pd.read_pickle('Data19.pkl')

After reopening the file in a new notebook, we have been able to perform all the commands we were willing to run with no issues at all. First of all, we are going to check the types of the variables.

df.dtypes

As we can see, Delay, Flights and Dist are floats which take a larger amount of memory than integers. Because we are working with a very large dataset, it is in our best interest to optimize the use of the memory throughout the process, and since we really don't lose any information by converting these variables into integers, this is our next step.

df['Flights'] = df['Flights'].astype('int32')

df['Dist'] = df['Dist'].astype('int32')

df['Delay'] = df['Delay'].astype('int32')

df

df.dtypes

### Grouping by Route, Month and Carrier.
After the dataframe has been joined, cleaned and we have all the information we want to use, we are going to perform the grouping.

Grouping by the forementioned categories allows us to get in a single observation the amount of flights performing each route in each month by each carrier. We also want to keep the rest of the columns, so we have to choose an aggregation method. For most of the features, they contain the same values: e.g. for all flights departing from ABE airport, FCity will be 'Allentown/Bethlehem/Easton, PA', while FST will be the departure state abbreviation, 'PA', and so on and so forth. 

Because of this, we are going to choose the aggregation method 'first' as it is the one that optimizes the memory usage: instead of making a calculation within all the possible options, it just grabs the value of the first observation and continues on with the next group. 

For the rest of the variables, which are the numeric ones, we are going to take the mean, except for the variable Flights where we are interested in the amount of flights, so we are just taking a count.

After we perform the grouping, the dataframe is going to be shown as a main category and sub-categories. In order to remove this and just get an unstructured dataframe, we will need to reset_index().

dfg = df.groupby(by=['Route', 'Month', 'Carrier']).agg({'From': 'first', 'FCity': 'first', 'FST': 'first', 
                                                        'To': 'first', 'TCity': 'first', 'TST': 'first',
                                                        'Delay': 'mean', 'Flights': 'sum', 'Dist': 'mean'})

dfg

dfg = dfg.reset_index()

dfg

Once again, due to technical difficulties, we are going to export the dataframe as a .pkl to continue working in another notebook.

dfg.to_pickle('GroupedData.pkl')
