==> 03.JoinStateInfo.ipynb <==
import pandas as pd

df = pd.read_pickle('GroupedData.pkl')

df

### Joining other data.
For the purpose of our research, we are interested in checking whether there is any correlation between some demographic and economic variables relative to each city or state and the amount of flights operating in such state. 

Our next steps are going to be obtaining the data for the state and city information we are willing to work with and join it to the main dataframe.

stdf = pd.read_excel('Data/StateData.xlsx')

stdf

stdf.columns = ['State', 'ST', 'Pop', 'GDP', 'Share', 'GDPpc']

stdf.drop(['Pop'], axis = 1, inplace = True)
stdf.drop(['GDP'], axis = 1, inplace = True)

stdf['Share'] = stdf['Share']*100

stdf[['Share', 'GDPpc']] = round(stdf[['Share', 'GDPpc']], 2)

df = df.merge(stdf, left_on = 'FST', right_on = 'ST', how = 'left')
df = df.merge(stdf, left_on = 'TST', right_on = 'ST', how = 'left')

df

df.columns

df.columns = ['Route', 'Month', 'Carrier', 'From', 'FCity', 'FST', 'To', 'TCity', 'TST', 
              'Delay', 'Flights', 'Dist', 'State_x', 'ST_x', 'FShare', 'FGDPpc',
              'State_y', 'ST_y', 'TShare', 'TGDPpc']

df = df[['Route', 'Month', 'Carrier',
          'From', 'FCity', 'FST', 'FShare', 'FGDPpc',
          'To', 'TCity', 'TST', 'TShare', 'TGDPpc',
          'Delay', 'Flights', 'Dist']]

df

We export this time to two different formats. As usual, .pkl will be the one we will be working in Python, and is our preferred format for storing the dataframe. However, we also want to use Tableau, and it will not read .pkl files, so we are also going to store the dataset as a .csv file so we are able to open it on the visualisation software.

df.to_pickle('FinalData.pkl')

df.to_csv('FinalData.csv')
