==> 04.JoinCityInfo.ipynb <==
import pandas as pd
from tqdm import tqdm
import numpy as np
import warnings
warnings.filterwarnings("ignore")

df = pd.read_pickle('FinalData.pkl')

Next step is for us to remove the last 4 characters on the city columns. Those 4 characters refer to the state they belong to, but that information is also stored in the adjacent columns, so we would like to get the clean name of the city in order to join the table with another table, where we will obtain the city population.

for i in tqdm(range(len(df))):
    df['FCity'].iloc[i] = df['FCity'].iloc[i][:-4]
    df['TCity'].iloc[i] = df['TCity'].iloc[i][:-4]

df

Finally, we are going to replace all the city fields that contain more than one city name. There are some airports that belong to two different cities, and this will be an issue when we try to join the tables. For that sole reason, we require city fields to contain the name of only one city.

df.FCity.replace('New York', 'New York City', inplace = True)
df.FCity.replace('Newark', 'New York City', inplace = True)
df.FCity.replace('Dallas/Fort Worth', 'Dallas', inplace = True)
df.FCity.replace('Montrose/Delta', 'Montrose', inplace = True)
df.FCity.replace('Newburgh/Poughkeepsie', 'New York City', inplace = True)
df.FCity.replace('Gulfport/Biloxi', 'Gulfport', inplace = True)
df.FCity.replace('Allentown/Bethlehem/Easton', 'Allentown', inplace = True)
df.FCity.replace('Raleigh/Durham', 'Raleigh', inplace = True)
df.FCity.replace('Jacksonville/Camp Lejeune', 'Jacksonville', inplace = True)
df.TCity.replace('New York', 'New York City', inplace = True)
df.TCity.replace('Newark', 'New York City', inplace = True)
df.TCity.replace('Dallas/Fort Worth', 'Dallas', inplace = True)
df.TCity.replace('Montrose/Delta', 'Montrose', inplace = True)
df.TCity.replace('Newburgh/Poughkeepsie', 'New York City', inplace = True)
df.TCity.replace('Gulfport/Biloxi', 'Gulfport', inplace = True)
df.TCity.replace('Allentown/Bethlehem/Easton', 'Allentown', inplace = True)
df.TCity.replace('Raleigh/Durham', 'Raleigh', inplace = True)
df.TCity.replace('Jacksonville/Camp Lejeune', 'Jacksonville', inplace = True)

df

Finally, we are going to merge the city population into the main dataframe. 

We have information on the top 200 cities in the US, which have a population of 140.000 or more. For all those cities not included in this list of 200, we are going to assign a population of 100.000 as a default. This will not be too important as we will see later how is population taken into account when we perform the model.

Moreover, we are going to work with the population in millions. For this reason, we also perform the calculation, and it makes working with this numbers easier because of their smaller scale.

citydf = pd.read_excel('Data/PopData.xlsx')
citydf['Pop'] = round(citydf['Pop']/1000000, 2)
citydf

df = df.merge(citydf, left_on = 'FCity', right_on = 'City', how = 'left')
df = df.merge(citydf, left_on = 'TCity', right_on = 'City', how = 'left')

df = df.drop(['City_x', 'City_y'], axis = 1)

df = df.fillna(0.1)

df

### Defining dummy variables.
We are now going to define several dummy variables. Dummies are binary variables that take either the value 0 or the value 1 depending on whether the observation does or does not satisfy a condition. These dummies are, by definition, creating two categories that are assumed to have differences due to the fact of belonging to such categories.

The first important feature we want to store as a dummy variable is whether the route is departing from or arriving to one of the US hubs. Hubs are main airports where the airlines set their operational bases. It is impossible for any airline to perform all possible routes within any pair of airports in the US. However, companies usually make all their destinations connect to their hubs, where then passengers can transit into another flight of the same company that completes the desired journey. 

Due to its nature, hubs are airports with a huge load of incoming and departing flights, and so it seems safe to assume that the amount of flights in a route will be heavily determined by whether one of the two connected airports is a hub or not.

On the other hand, we also want to take a look at flights going to or departing from Hawaii and Alaska. Those two states are the ones which are not part of mainland US, and the only way to connect them to the rest of states is by an air connection. We can also assume that, given there are no alternatives to travel to or from those states, there is going to be a larger number of flights for such routes, and then, it will be a factor we should take into account. 

By the same logic, we can also create a feature that will respond to whether the route is connecting two cities within the same state. There are some routes that connect a pair of cities of the same state, and in this case, there may be other means of transportation that could be used other than planes. For this reason, we also believe this could be a factor that influences the amount of flights in a route, so we will also create such a dummy.

hubs = ['ATL', 'LAX', 'DFW', 'EWR', 'JFK', 'SFO', 'DEN', 'MIA', 'ORD', 'BOS', 'DAL', 'SEA']
away = ['HI', 'AK']

df['Hub'] = np.where(df['From'].isin(hubs)+df['To'].isin(hubs)>0, 1, 0)

df['Hub'].value_counts()

df['Away'] = np.where(df['FST'].isin(away)+df['TST'].isin(away)>0, 1, 0)

df['Away'].value_counts()

df['Within'] = np.where(df['FST'] == df['TST'], 1, 0)

df

df['Delay'] = round(df['Delay'], 0)

df

df.to_csv('Flights19.csv')

df.to_pickle('Flights19.pkl')
