==> 05.CreateTableauDF.ipynb <==
import pandas as pd

Finally, we are going to join the data we have on the airports with geolocation data for each airport. This is a necessary step to work accurately with Tableau.

For our visualisations, we are not going to use all the features we have or we created, so for that reason, we are going to take the least amount of columns only to avoid having an unnecessarily large file.

dft = pd.read_csv('Flights19.csv')

air = pd.read_excel('Data/Airports.xlsx')

dft = dft[['Route', 'Month', 'Carrier', 'From', 'FST', 'FShare', 'FGDPpc', 'To', 'Delay', 'Flights']]

air

dft = pd.merge(dft, air, left_on = 'From', right_on = 'Airport')

dft = pd.merge(dft, air, left_on = 'To', right_on = 'Airport')

dft = dft[['Route', 'Month', 'Carrier', 'From', 'FST', 'FShare', 'FGDPpc', 'To',
       'Delay', 'Flights', 'Lat_x', 'Long_x', 
       'Lat_y', 'Long_y']]

dft.columns = [['Route', 'Month', 'Carrier', 'From', 'FST', 'FShare', 'FGDPpc', 'To',
       'Delay', 'Flights', 'Lat_x', 'Long_x', 'Lat_y', 'Long_y']]

dft.to_csv('TabDF.csv')
