==> 06.FindingRouteOpportinities.ipynb <==
import pandas as pd
import warnings
warnings.filterwarnings("ignore")

### Finding Route Opportunities
In this notebook we are going to finally get some insights about the data by answering one of the important questions that drives our research: are there any pairs of cities that do not have a direct route connecting them? If the answer is yes, then we could be pointing at a market opportunity for an airline company, and this would lead to the second part of our research: how many flights could populate this route according to the US aviation industry?

To perform this first step, we are going to take a list with the 40 busiest airports in the US and we are going to pair them with each other and see whether this combination exists in our dataset. If the combination exists, there is already (at least) a carrier performing this route, and thus it is not an opportunity. Then, we are going to be interested in the pairs of airports that are not in our dataset. In that case, there are no carriers performing this route and it could be the opportunity we are looking for.

df = pd.read_pickle('Flights19.pkl')
df

df

# First, we obtain the top 40 busiest airports in the US
# by the total amount of flights arriving and departing from them.

top40 = df.groupby('From').agg({'Flights': 'sum'}).sort_values('Flights', ascending = False).reset_index()['From'][:40]

top40

# Secondly we take the list of the 40 airports and create 40x39 pairs of airports and
# check whether each of the combinations is in the main dataframe.
# We print out the combinations that do not exist in the dataframe.

for i in range(40):
    for j in range(i+1,40):
        route = top40[i] + '-' + top40[j]
        if route not in list(df['Route']):
            print(route)

We can see that there are a lot of possible routes to explore. However, there is something we must take into account. There are some cities which are served by more than one airport. So a route from Charlotte to New York City may exist, but it will surely not exist for all 3 airports in New York, so our previous algorithm will return the other two airports as an opportunity, while it is not.

To solve this issue, we are going to group all major airports that are within 2h drive of each other, as they serve the same region. We will not be comparing pairs of airports anymore, but we will be comparing pairs of unconnected cities.

df2 = df.reset_index(drop = True)

df2['From'].replace('JFK', 'NYC', True)
df2['From'].replace('EWR', 'NYC', True)
df2['From'].replace('LGA', 'NYC', True)
df2['From'].replace('ORD', 'CHI', True)
df2['From'].replace('MDW', 'CHI', True)
df2['From'].replace('MCO', 'TPA', True)
df2['From'].replace('DFW', 'DAL', True)
df2['From'].replace('PBI', 'MIA', True)
df2['From'].replace('FLL', 'MIA', True)
df2['From'].replace('SAN', 'LAX', True)
df2['From'].replace('ONT', 'LAX', True)
df2['From'].replace('SNA', 'LAX', True)
df2['From'].replace('IAD', 'WAS', True)
df2['From'].replace('DCA', 'WAS', True)
df2['From'].replace('IAH', 'HOU', True)
df2['From'].replace('SJC', 'SFO', True)
df2['From'].replace('OAK', 'SFO', True)
df2['From'].replace('SMF', 'SFO', True)

df2['To'].replace('JFK', 'NYC', True)
df2['To'].replace('EWR', 'NYC', True)
df2['To'].replace('LGA', 'NYC', True)
df2['To'].replace('ORD', 'CHI', True)
df2['To'].replace('MDW', 'CHI', True)
df2['To'].replace('MCO', 'TPA', True)
df2['To'].replace('DFW', 'DAL', True)
df2['To'].replace('PBI', 'MIA', True)
df2['To'].replace('FLL', 'MIA', True)
df2['To'].replace('SAN', 'LAX', True)
df2['To'].replace('ONT', 'LAX', True)
df2['To'].replace('SNA', 'LAX', True)
df2['To'].replace('IAD', 'WAS', True)
df2['To'].replace('DCA', 'WAS', True)
df2['To'].replace('IAH', 'HOU', True)
df2['To'].replace('SJC', 'SFO', True)
df2['To'].replace('OAK', 'SFO', True)
df2['To'].replace('SMF', 'SFO', True)

df2

# After grouping the airports by city, we can see now that the busiest city is New York City if you aggregate
# the amount of flights on the 3 airports serving the city. We are going to perform the same steps as before, 
# taking into account that we are now going to compare pairs of cities.

top35 = df2.groupby('From').agg({'Flights': 'sum'}).sort_values('Flights', ascending = False).reset_index()['From'][:35]

top35

df2['Route'] = df2['From']+'-'+df2['To']

for i in range(35):
    for j in range(i+1,35):
        route = top35[i] + '-' + top35[j]
        if route not in list(df2['Route']):
            print(route)

The final list would be the list of opportunities: pairs of unconnected cities that could potentially be a new route. We are going to choose 3 of them based on population. For this to be a relevant research, we are willing to find big cities that are not yet connected, rather than smaller ones which may not have a significant demand for a flight to other cities.
